+++ concordium-std/src/impls.rs
@@ -1,4 +1,5 @@
 use crate::{
+    boxed,
     cell::UnsafeCell,
     collections::{BTreeMap, BTreeSet},
     convert::{self, TryInto},
@@ -2120,6 +2121,69 @@ impl HasHost<ExternStateApi> for ExternLowLevelHost {
     }
 }
 
+impl HasCryptoPrimitives for ExternCryptoPrimitives {
+    fn verify_ed25519_signature(
+        &self,
+        public_key: PublicKeyEd25519,
+        signature: SignatureEd25519,
+        message: &[u8],
+    ) -> bool {
+        let res = unsafe {
+            prims::verify_ed25519_signature(
+                public_key.0.as_ptr(),
+                signature.0.as_ptr(),
+                message.as_ptr(),
+                message.len() as u32,
+            )
+        };
+        res == 1
+    }
+
+    fn verify_ecdsa_secp256k1_signature(
+        &self,
+        public_key: PublicKeyEcdsaSecp256k1,
+        signature: SignatureEcdsaSecp256k1,
+        message_hash: [u8; 32],
+    ) -> bool {
+        let res = unsafe {
+            prims::verify_ecdsa_secp256k1_signature(
+                public_key.0.as_ptr(),
+                signature.0.as_ptr(),
+                message_hash.as_ptr(),
+            )
+        };
+        res == 1
+    }
+
+    fn hash_sha2_256(&self, data: &[u8]) -> Hash256 {
+        let mut output: MaybeUninit<[u8; 32]> = MaybeUninit::uninit();
+        unsafe {
+            prims::hash_sha2_256(data.as_ptr(), data.len() as u32, output.as_mut_ptr() as *mut u8);
+            Hash256(output.assume_init())
+        }
+    }
+
+    fn hash_sha3_256(&self, data: &[u8]) -> Hash256 {
+        let mut output: MaybeUninit<[u8; 32]> = MaybeUninit::uninit();
+        unsafe {
+            prims::hash_sha3_256(data.as_ptr(), data.len() as u32, output.as_mut_ptr() as *mut u8);
+            Hash256(output.assume_init())
+        }
+    }
+
+    fn hash_keccak_256(&self, data: &[u8]) -> Hash256 {
+        let mut output: MaybeUninit<[u8; 32]> = MaybeUninit::uninit();
+        unsafe {
+            prims::hash_keccak_256(
+                data.as_ptr(),
+                data.len() as u32,
+                output.as_mut_ptr() as *mut u8,
+            );
+            Hash256(output.assume_init())
+        }
+    }
+}
+
 /// # Trait implementations for the init context
 impl HasInitContext for ExternContext<crate::types::ExternInitContext> {
     type InitData = ();
@@ -2602,4 +2666,84 @@ where
     V: Serial + DeserialWithState<S> + Deletable,
 {
     fn delete(mut self) { self.clear(); }
+}
+
+impl Serial for PublicKeyEd25519 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.0.serial(out) }
+}
+
+impl Deserial for PublicKeyEd25519 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        Ok(PublicKeyEd25519(Deserial::deserial(source)?))
+    }
+}
+
+impl schema::SchemaType for PublicKeyEd25519 {
+    fn get_type() -> concordium_contracts_common::schema::Type {
+        schema::Type::Array(32, boxed::Box::new(schema::Type::U8))
+    }
+}
+
+impl Serial for PublicKeyEcdsaSecp256k1 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.0.serial(out) }
+}
+
+impl Deserial for PublicKeyEcdsaSecp256k1 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        Ok(PublicKeyEcdsaSecp256k1(Deserial::deserial(source)?))
+    }
+}
+
+impl schema::SchemaType for PublicKeyEcdsaSecp256k1 {
+    fn get_type() -> concordium_contracts_common::schema::Type {
+        schema::Type::Array(33, boxed::Box::new(schema::Type::U8))
+    }
+}
+
+impl Serial for SignatureEd25519 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.0.serial(out) }
+}
+
+impl Deserial for SignatureEd25519 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        Ok(SignatureEd25519(Deserial::deserial(source)?))
+    }
+}
+
+impl schema::SchemaType for SignatureEd25519 {
+    fn get_type() -> concordium_contracts_common::schema::Type {
+        schema::Type::Array(64, boxed::Box::new(schema::Type::U8))
+    }
+}
+
+impl Serial for SignatureEcdsaSecp256k1 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.0.serial(out) }
+}
+
+impl Deserial for SignatureEcdsaSecp256k1 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        Ok(SignatureEcdsaSecp256k1(Deserial::deserial(source)?))
+    }
+}
+
+impl schema::SchemaType for SignatureEcdsaSecp256k1 {
+    fn get_type() -> concordium_contracts_common::schema::Type {
+        schema::Type::Array(64, boxed::Box::new(schema::Type::U8))
+    }
+}
+
+impl Serial for Hash256 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.0.serial(out) }
+}
+
+impl Deserial for Hash256 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        Ok(Hash256(Deserial::deserial(source)?))
+    }
+}
+
+impl schema::SchemaType for Hash256 {
+    fn get_type() -> concordium_contracts_common::schema::Type {
+        schema::Type::Array(32, boxed::Box::new(schema::Type::U8))
+    }
 }