+++ concordium-std/src/traits.rs
@@ -6,8 +6,9 @@
 use crate::vec::Vec;
 use crate::{
     types::{LogError, StateError},
-    CallContractResult, EntryRaw, Key, OccupiedEntryRaw, ReadOnlyCallContractResult, StateBuilder,
-    TransferResult, VacantEntryRaw,
+    CallContractResult, EntryRaw, Hash256, Key, OccupiedEntryRaw, PublicKeyEcdsaSecp256k1,
+    PublicKeyEd25519, ReadOnlyCallContractResult, SignatureEcdsaSecp256k1, SignatureEd25519,
+    StateBuilder, TransferResult, VacantEntryRaw,
 };
 use concordium_contracts_common::*;
 
@@ -402,6 +403,35 @@ pub trait HasLogger {
         }
         self.log_raw(&out)
     }
+}
+
+/// Objects which provide cryptographic primitives.
+pub trait HasCryptoPrimitives {
+    /// Verify an ed25519 signature.
+    fn verify_ed25519_signature(
+        &self,
+        public_key: PublicKeyEd25519,
+        signature: SignatureEd25519,
+        message: &[u8],
+    ) -> bool;
+
+    /// Verify an ecdsa signature over secp256k1 with the bitcoin-core
+    /// implementation.
+    fn verify_ecdsa_secp256k1_signature(
+        &self,
+        public_key: PublicKeyEcdsaSecp256k1,
+        signature: SignatureEcdsaSecp256k1,
+        message_hash: [u8; 32],
+    ) -> bool;
+
+    /// Hash the data using the SHA2-256 algorithm.
+    fn hash_sha2_256(&self, data: &[u8]) -> Hash256;
+
+    /// Hash the data using the SHA3-256 algorithm.
+    fn hash_sha3_256(&self, data: &[u8]) -> Hash256;
+
+    /// Hash the data using the Keccak-256 algorithm.
+    fn hash_keccak_256(&self, data: &[u8]) -> Hash256;
 }
 
 /// Add optimized unwrap behaviour that aborts the process instead of