+++ concordium-std/src/types.rs
@@ -954,6 +954,37 @@ impl<S> StateBuilder<S> {
     pub fn into_inner(self) -> S { self.state_api }
 }
 
+/// A struct for which HasCryptoPrimitives is implemented via the crypto host
+/// functions.
+#[doc(hidden)]
+pub struct ExternCryptoPrimitives;
+
+/// Public key for Ed25519. Must be 32 bytes long.
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
+#[repr(transparent)]
+pub struct PublicKeyEd25519(pub [u8; 32]);
+
+/// Public key for ECDSA over Secp256k1. Must be 33 bytes long.
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
+#[repr(transparent)]
+pub struct PublicKeyEcdsaSecp256k1(pub [u8; 33]);
+
+/// Signature for a Ed25519 message. Must be 64 bytes long.
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
+#[repr(transparent)]
+pub struct SignatureEd25519(pub [u8; 64]);
+
+/// Signature for a ECDSA (over Secp256k1) message. Must be 64 bytes longs
+/// (serialized in compressed format).
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
+#[repr(transparent)]
+pub struct SignatureEcdsaSecp256k1(pub [u8; 64]);
+
+/// Hash digest with 256 bits (32 bytes).
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
+#[repr(transparent)]
+pub struct Hash256(pub [u8; 32]);
+
 #[derive(Debug, Clone, Default)]
 #[doc(hidden)]
 pub struct ExternStateApi;